+++ Urasandesu.Swathe/Urasandesu/Swathe/Metadata/BaseClass/BaseMethodMetadata.hpp
@@ -199,7 +199,7 @@ namespace Urasandesu { namespace Swathe { namespace Metadata { namespace BaseCla
     }
 
     template<class ApiHolder>    
-    ULONG BaseMethodMetadata<ApiHolder>::GetHashCode() const
+    size_t BaseMethodMetadata<ApiHolder>::GetHashCode() const
     {
         return Pimpl()->GetHashCode();
     }
