+++ Urasandesu.Swathe/Urasandesu/Swathe/Metadata/BaseClass/BaseParameterGenerator.hpp
@@ -151,7 +151,7 @@ namespace Urasandesu { namespace Swathe { namespace Metadata { namespace BaseCla
     }
 
     template<class ApiHolder>    
-    ULONG BaseParameterGenerator<ApiHolder>::GetHashCode() const
+    size_t BaseParameterGenerator<ApiHolder>::GetHashCode() const
     {
         return Pimpl()->GetHashCode();
     }
