+++ Urasandesu.Swathe/Urasandesu/Swathe/Metadata/BaseClassPimpl/BaseMethodBodyGeneratorPimpl.h
@@ -126,7 +126,7 @@ namespace Urasandesu { namespace Swathe { namespace Metadata { namespace BaseCla
         IAssembly const *GetAssembly() const;
         IMethodBody const *GetSourceMethodBody() const;
         bool Equals(IMethodBody const *pBody) const;
-        ULONG GetHashCode() const;
+        size_t GetHashCode() const;
         ILocal const *GetLocal(ULONG index, IType const *pLocalType) const;
         IInstructionPtrRange::iterator GetInstructionIterator(ULONG offset) const;
         void Emit(OpCode const &opCode, ILocal const *pLocal);
